% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_HAIgermany_app}
\alias{run_HAIgermany_app}
\title{Launch the HAIgermany Shiny App}
\usage{
run_HAIgermany_app()
}
\value{
Runs the Shiny app in a new window.
}
\description{
This function launches the interactive Shiny app included in the HAIgermany package.
}
\examples{
if (interactive()) {
  run_HAIgermany_app()
}
}
